+++ release/installssh.sh
@@ -6,11 +6,6 @@ CHROOTDIR=~/cos/mkiso_out/squashfs-root
 DEBDIR=~/app
 DEBNAME=ssh.tar.gz
 
-if [ ! -e "${CHROOTDIR}" ]; then
-    echo "squashfs-root not found"
-    exit
-fi
-
 if [ -z "$1" ] ; then
     echo error: No chrootdir setting at first param.
     exit -1
@@ -24,6 +19,10 @@ fi
 CHROOTDIR=$1/squashfs-root
 DEBDIR=$2
 
+if [ ! -e "${CHROOTDIR}" ]; then
+    echo "squashfs-root not found"
+    exit
+fi
 
 mkdir ${CHROOTDIR}/app
 sudo cp ${DEBDIR}/${DEBNAME} ${CHROOTDIR}/app
